requirements.txt
requests>=2.31.0
Pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.7.0
